+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/parser/FilterObjectToMap.kt
@@ -19,7 +19,7 @@ import io.getstream.chat.android.client.api.models.NotExistsFilterObject
 import io.getstream.chat.android.client.api.models.NotInFilterObject
 import io.getstream.chat.android.client.api.models.OrFilterObject
 
-internal fun FilterObject.toMap(): Map<*, *> = when (this) {
+internal fun FilterObject.toMap(): Map<String, Any> = when (this) {
     is AndFilterObject -> mapOf(KEY_AND to this.filterObjects.map(FilterObject::toMap))
     is OrFilterObject -> mapOf(KEY_OR to this.filterObjects.map(FilterObject::toMap))
     is NorFilterObject -> mapOf(KEY_NOR to this.filterObjects.map(FilterObject::toMap))