+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/input/MessageInputView.kt
@@ -22,9 +22,11 @@ import com.getstream.sdk.chat.utils.extensions.activity
 import com.getstream.sdk.chat.utils.extensions.focusAndShowKeyboard
 import com.google.android.material.snackbar.Snackbar
 import io.getstream.chat.android.client.logger.ChatLogger
+import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Command
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
+import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
 import io.getstream.chat.android.ui.R
 import io.getstream.chat.android.ui.common.Debouncer
@@ -34,10 +36,12 @@ import io.getstream.chat.android.ui.common.extensions.internal.streamThemeInflat
 import io.getstream.chat.android.ui.common.style.setTextStyle
 import io.getstream.chat.android.ui.databinding.StreamUiMessageInputBinding
 import io.getstream.chat.android.ui.message.input.MessageInputView.MaxMessageLengthHandler
+import io.getstream.chat.android.ui.message.input.MessageInputView.MessageInputViewModeListener
 import io.getstream.chat.android.ui.message.input.MessageInputView.SelectedAttachmentsCountListener
 import io.getstream.chat.android.ui.message.input.attachment.AttachmentSelectionDialogFragment
 import io.getstream.chat.android.ui.message.input.attachment.AttachmentSelectionListener
 import io.getstream.chat.android.ui.message.input.attachment.AttachmentSource
+import io.getstream.chat.android.ui.message.input.attachment.selected.internal.SelectedCustomAttachmentViewHolderFactory
 import io.getstream.chat.android.ui.message.input.internal.MessageInputFieldView
 import io.getstream.chat.android.ui.message.input.mention.searchUsers
 import io.getstream.chat.android.ui.message.input.transliteration.DefaultStreamTransliterator
@@ -108,6 +112,12 @@ public class MessageInputView : ConstraintLayout {
             }
         }
     }
+
+    private val customAttachmentsSelectionListener =
+        { attachments: Collection<Attachment>, viewHolderFactory: SelectedCustomAttachmentViewHolderFactory ->
+            binding.messageInputFieldView.mode =
+                MessageInputFieldView.Mode.CustomAttachmentMode(attachments.toList(), viewHolderFactory)
+        }
 
     /**
      * The default implementation of [MaxMessageLengthHandler] which uses the default [EditText] error
@@ -605,6 +615,10 @@ public class MessageInputView : ConstraintLayout {
                 override fun onModeChanged(mode: MessageInputFieldView.Mode) {
                     refreshControlsState()
                 }
+
+                override fun onSelectedCustomAttachmentsChanged(selectedCustomAttachments: List<Attachment>) {
+                    refreshControlsState()
+                }
             }
         )
 
@@ -711,43 +725,68 @@ public class MessageInputView : ConstraintLayout {
     }
 
     private fun sendMessage(messageReplyTo: Message? = null) {
+        val messageText = binding.messageInputFieldView.messageText
+
         doSend(
             { attachments ->
                 sendMessageHandler.sendMessageWithAttachments(
-                    binding.messageInputFieldView.messageText,
+                    messageText,
                     attachments,
                     messageReplyTo
                 )
             },
-            { sendMessageHandler.sendMessage(binding.messageInputFieldView.messageText, messageReplyTo) }
+            { sendMessageHandler.sendMessage(messageText, messageReplyTo) },
+            { customAttachments ->
+                sendMessageHandler.sendMessageWithCustomAttachments(messageText, customAttachments, messageReplyTo)
+            }
         )
     }
 
     private fun sendThreadMessage(parentMessage: Message) {
         val sendAlsoToChannel = binding.sendAlsoToChannel.isChecked
-        doSend({ attachments ->
-            sendMessageHandler.sendToThreadWithAttachments(
-                parentMessage,
-                binding.messageInputFieldView.messageText,
-                sendAlsoToChannel,
-                attachments
-            )
-        },
+        val messageText = binding.messageInputFieldView.messageText
+        doSend(
+            { attachments ->
+                sendMessageHandler.sendToThreadWithAttachments(
+                    parentMessage,
+                    messageText,
+                    sendAlsoToChannel,
+                    attachments
+                )
+            },
             {
                 sendMessageHandler.sendToThread(parentMessage,
-                    binding.messageInputFieldView.messageText,
+                    messageText,
                     sendAlsoToChannel)
+            },
+            { customAttachments ->
+                sendMessageHandler.sendToThreadWithCustomAttachments(
+                    parentMessage,
+                    messageText,
+                    sendAlsoToChannel,
+                    customAttachments)
             }
         )
     }
 
-    private fun doSend(attachmentSender: (List<Pair<File, String?>>) -> Unit, simpleSender: () -> Unit) {
+    private fun doSend(
+        attachmentSender: (List<Pair<File, String?>>) -> Unit,
+        simpleSender: () -> Unit,
+        customAttachmentsSender: (List<Attachment>) -> Unit,
+    ) {
         val attachments = binding.messageInputFieldView.getAttachedFiles()
+        val customAttachments = binding.messageInputFieldView.getCustomAttachments()
 
-        if (attachments.isNotEmpty()) {
-            attachmentSender(attachments)
-        } else {
-            simpleSender()
+        when {
+            attachments.isNotEmpty() -> {
+                attachmentSender(attachments)
+            }
+            customAttachments.isNotEmpty() -> {
+                customAttachmentsSender(customAttachments)
+            }
+            else -> {
+                simpleSender()
+            }
         }
     }
 
@@ -766,6 +805,19 @@ public class MessageInputView : ConstraintLayout {
         attachmentSelectionListener.onAttachmentsSelected(attachments.toSet(), attachmentSource)
     }
 
+    /**
+     * Set a collection of custom attachments in [MessageInputView].
+     *
+     * @param attachments Collection of [Attachment] that you are going to send with a message.
+     */
+    @ExperimentalStreamChatApi
+    public fun submitCustomAttachments(
+        attachments: Collection<Attachment>,
+        viewHolderFactory: SelectedCustomAttachmentViewHolderFactory,
+    ) {
+        customAttachmentsSelectionListener(attachments, viewHolderFactory)
+    }
+
     private companion object {
         private const val CLICK_DELAY = 100L
         private const val TYPING_DEBOUNCE_MS = 300L
@@ -780,6 +832,14 @@ public class MessageInputView : ConstraintLayout {
                 messageReplyTo: Message?,
             ) {
                 throw IllegalStateException("MessageInputView#messageSendHandler needs to be configured to send messages")
+            }
+
+            override fun sendMessageWithCustomAttachments(
+                message: String,
+                attachments: List<Attachment>,
+                messageReplyTo: Message?,
+            ) {
+                throw IllegalStateException("MessageInputView#sendMessageWithCustomAttachments needs to be configured to send messages")
             }
 
             override fun sendToThread(
@@ -797,6 +857,15 @@ public class MessageInputView : ConstraintLayout {
                 attachmentsWithMimeTypes: List<Pair<File, String?>>,
             ) {
                 throw IllegalStateException("MessageInputView#messageSendHandler needs to be configured to send messages")
+            }
+
+            override fun sendToThreadWithCustomAttachments(
+                parentMessage: Message,
+                message: String,
+                alsoSendToChannel: Boolean,
+                attachments: List<Attachment>,
+            ) {
+                throw IllegalStateException("MessageInputView#sendToThreadWithCustomAttachments needs to be configured to send messages")
             }
 
             override fun editMessage(oldMessage: Message, newMessageText: String) {
@@ -851,6 +920,12 @@ public class MessageInputView : ConstraintLayout {
             messageReplyTo: Message? = null,
         )
 
+        public fun sendMessageWithCustomAttachments(
+            message: String,
+            attachments: List<Attachment>,
+            messageReplyTo: Message? = null,
+        )
+
         public fun sendToThread(
             parentMessage: Message,
             messageText: String,
@@ -862,6 +937,13 @@ public class MessageInputView : ConstraintLayout {
             message: String,
             alsoSendToChannel: Boolean,
             attachmentsWithMimeTypes: List<Pair<File, String?>>,
+        )
+
+        public fun sendToThreadWithCustomAttachments(
+            parentMessage: Message,
+            message: String,
+            alsoSendToChannel: Boolean,
+            attachmentsWithMimeTypes: List<Attachment>,
         )
 
         public fun editMessage(oldMessage: Message, newMessageText: String)