+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/MessageListView.kt
@@ -266,7 +266,6 @@ public class MessageListView : ConstraintLayout {
     private var messageListItemPredicate: MessageListItemPredicate = HiddenMessageListItemPredicate
     private var messageListItemTransformer: MessageListItemTransformer = MessageListItemTransformer { it }
     private var showAvatarPredicate: ShowAvatarPredicate = DefaultShowAvatarPredicate()
-    private var messageBackgroundFactory: MessageBackgroundFactory? = null
 
     private var deletedMessageListItemPredicate: MessageListItemPredicate =
         DeletedMessageListItemPredicate.VisibleToEveryone
@@ -299,7 +298,7 @@ public class MessageListView : ConstraintLayout {
                         ),
                         requireStyle(),
                         messageListItemViewHolderFactory,
-                        backgroundFactory()
+                        messageBackgroundFactory
                     )
                     .apply {
                         setReactionClickHandler { message, reactionType ->
@@ -401,7 +400,7 @@ public class MessageListView : ConstraintLayout {
                     ),
                     requireStyle(),
                     messageListItemViewHolderFactory,
-                    backgroundFactory()
+                    messageBackgroundFactory
                 ).apply {
                     setReactionClickHandler { message, reactionType ->
                         messageReactionHandler.onMessageReaction(message, reactionType)
@@ -447,6 +446,7 @@ public class MessageListView : ConstraintLayout {
     private lateinit var messageDateFormatter: DateFormatter
     private lateinit var attachmentViewFactory: AttachmentViewFactory
     private lateinit var attachmentViewHolderFactory: AttachmentViewHolderFactory
+    private lateinit var messageBackgroundFactory: MessageBackgroundFactory
 
     public constructor(context: Context) : this(context, null, 0)
     public constructor(context: Context, attrs: AttributeSet?) : this(context, attrs, 0)
@@ -489,10 +489,6 @@ public class MessageListView : ConstraintLayout {
                 loadingViewContainer.addView(this)
             }
         }
-    }
-
-    private fun backgroundFactory(): MessageBackgroundFactory {
-        return messageBackgroundFactory ?: MessageBackgroundFactoryImpl(requireStyle().itemStyle)
     }
 
     private fun initEmptyStateView() {
@@ -651,12 +647,16 @@ public class MessageListView : ConstraintLayout {
             messageListItemViewHolderFactory = MessageListItemViewHolderFactory()
         }
 
+        if (::messageBackgroundFactory.isInitialized.not()) {
+            messageBackgroundFactory = MessageBackgroundFactoryImpl(requireStyle().itemStyle)
+        }
+
         messageListItemViewHolderFactory.decoratorProvider = MessageListItemDecoratorProvider(
             dateFormatter = messageDateFormatter,
             isDirectMessage = { channel.isDirectMessaging() },
             messageListViewStyle = requireStyle(),
             showAvatarPredicate = this.showAvatarPredicate,
-            backgroundFactory()
+            messageBackgroundFactory
         )
 
         messageListItemViewHolderFactory.setListenerContainer(this.listenerContainer)
@@ -871,6 +871,19 @@ public class MessageListView : ConstraintLayout {
             "Adapter was already initialized, please set MessageViewHolderFactory first"
         }
         this.messageListItemViewHolderFactory = messageListItemViewHolderFactory
+    }
+
+    /**
+     * Allows clients to set a custom implementation of [MessageBackgroundFactory]. Use this
+     * method if you want to change the background of messages
+     *
+     * @param messageBackgroundFactory The custom factory that provides drawables to be used in the messages background
+     */
+    public fun setMessageBackgroundFactory(messageBackgroundFactory: MessageBackgroundFactory) {
+        check(::adapter.isInitialized.not()) {
+            "Adapter was already initialized, please set MessageViewHolderFactory first"
+        }
+        this.messageBackgroundFactory = messageBackgroundFactory
     }
 
     /**