+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/MessageListItemViewHolderFactory.kt
@@ -44,7 +44,7 @@ public open class MessageListItemViewHolderFactory {
      * Listeners set on [MessageListView] that should be invoked when the user interacts with
      * list items.
      */
-    protected lateinit var listenerContainer: MessageListListenerContainer
+    protected var listenerContainer: MessageListListenerContainer? = null
         private set
 
     /**
@@ -69,7 +69,7 @@ public open class MessageListItemViewHolderFactory {
 
     private val attachmentsRecycledViewPool: RecyclerView.RecycledViewPool = RecyclerView.RecycledViewPool()
 
-    internal fun setListenerContainer(listenerContainer: MessageListListenerContainer) {
+    internal fun setListenerContainer(listenerContainer: MessageListListenerContainer?) {
         this.listenerContainer = listenerContainer
     }
 
@@ -102,13 +102,26 @@ public open class MessageListItemViewHolderFactory {
     internal fun clone(): MessageListItemViewHolderFactory {
         val newFactory = MessageListItemViewHolderFactory()
 
-        if (::decoratorProvider.isInitialized) { newFactory.decoratorProvider = decoratorProvider }
-        if (::listenerContainer.isInitialized) { newFactory.listenerContainer = listenerContainer }
-        if (attachmentViewFactory != null) { newFactory.attachmentViewFactory = attachmentViewFactory }
-        if (attachmentViewHolderFactory != null) { newFactory.attachmentViewHolderFactory = attachmentViewHolderFactory }
-        if (::style.isInitialized) { newFactory.style = style }
-        if (::messageReplyStyle.isInitialized) { newFactory.messageReplyStyle = messageReplyStyle }
-        if (::giphyViewHolderStyle.isInitialized) { newFactory.giphyViewHolderStyle = giphyViewHolderStyle }
+        if (::decoratorProvider.isInitialized) {
+            newFactory.decoratorProvider = decoratorProvider
+        }
+        if (attachmentViewFactory != null) {
+            newFactory.attachmentViewFactory = attachmentViewFactory
+        }
+        if (attachmentViewHolderFactory != null) {
+            newFactory.attachmentViewHolderFactory = attachmentViewHolderFactory
+        }
+        if (::style.isInitialized) {
+            newFactory.style = style
+        }
+        if (::messageReplyStyle.isInitialized) {
+            newFactory.messageReplyStyle = messageReplyStyle
+        }
+        if (::giphyViewHolderStyle.isInitialized) {
+            newFactory.giphyViewHolderStyle = giphyViewHolderStyle
+        }
+
+        newFactory.listenerContainer = listenerContainer
 
         return newFactory
     }
@@ -192,7 +205,10 @@ public open class MessageListItemViewHolderFactory {
     protected fun createGiphyMessageItemViewHolder(
         parentView: ViewGroup,
     ): BaseMessageItemViewHolder<MessageListItem.MessageItem> {
-        return GiphyViewHolder(parentView, decoratorProvider.decorators, listenerContainer, giphyViewHolderStyle)
+        return GiphyViewHolder(parentView,
+            decoratorProvider.decorators,
+            listenerContainer,
+            giphyViewHolderStyle)
     }
 
     protected fun createSystemMessageItemViewHolder(